Optimize score calculation in Maths script
This commit replaces Lists with integer variables in the CalculateScienceScore function of the Maths.cs script, enhancing efficiency. It removes the need for repeatedly adding to and extracting max from a list, thus optimizing the calculation of science scores. It also adjusts related settings in workspace.xml and CalculatorInterface.tscn.