Add PerformanceTester and implement Reset function
This commit introduces a new script, PerformanceTester, that calculates the performance of the system. The Reset function is also added to the ScienceContainer and Main scripts, allowing for the resetting of certain values to their initial state. Minor modifications have been made to other files, mostly involving the rearrangement of existing methods.